Move keywords to Salesforce.py
