+++ cumulusci/robotframework/tests/cumulusci/base.robot
@@ -39,40 +39,3 @@ Test Run Task Class
 Test Run Task Class With Options
     Run Task Class  cumulusci.tasks.salesforce.CreatePackage  package=Test Package
 
-Test Perf Set Elapsed Time
-    [Tags]  perf
-    Set Test Elapsed Time       11655.9   #  3:14:15.9
-
-Test Perf Set Elapsed Time Twice
-    [Tags]  perf
-    Set Test Elapsed Time       11655.9
-    Set Test Elapsed Time       53  
-
-Test Perf Set Elapsed Time String
-    [Tags]  perf
-    Log                 A
-    Set Test Elapsed Time       5 hours   #  5:00:00
-
-Test Perf Measure Elapsed
-    [Setup]       Log             Before
-    [Teardown]    Log             After
-    Log                 B
-    Start Performance Timer
-    Sleep       1
-    Log         Noop
-    Stop Performance Timer
-
-Set Time and Also Metric
-    Start Performance Timer
-    Log         Noop
-    Stop Performance Timer
-    Set Test Metric   number of records  100 
-
-Test Perf Measure Other Metric
-    Set Test Metric    Max_CPU_Percent    30
-
-Mismatched Stop Performance Timer
-    Run Keyword and Expect Error
-    ...         *Elapsed time clock was not*
-    ...         Stop Performance Timer
-
